Add necessary message to pump script with warning that compilations were forced from distcc-pump mode into plain mode.
I had resisted doing this at first, because of the
max_discrepancies_before_demotion constant in src/compile.c, which actually
controls whether there were sufficiently many discrepancies to force a demotion.
The problem is that the constant is not known to the pump script. It should be
made configurable, through gosh, another env var.  But that'll be for later.

This constant is currently 1 and it shouldn't be changed.

Reviewers: fergus and csilvers.