Ensure that all ".c" files start with "#include <config.h>".
This will hopefully fix issue 39 <http://code.google.com/p/distcc/issues/detail?id=39>.
Some header files were using #if statements that depended on macros defined in config.h
without #including config.h first.  The result was a compilation error, because
"#if !HAVE_DECL_SNPRINTF" was evaluating to true because config.h hadn't been included,
causing us to declare snprintf() when we shouldn't, leading to a conflict with the
declaration in the standard header files.

(The alternative solution would have been to add "#include <config.h>"
to every header file that uses '#if'.  I decided not to do that, just
for consistency with the existing code.)